+++ pages/_error.tsx
@@ -1,0 +1,10 @@
+import { NextPage } from 'next'
+import { GenericError } from 'components/GenericError'
+
+type Props = {}
+
+const ErrorPage: NextPage<Props> = () => {
+  return <GenericError />
+}
+
+export default ErrorPage
