+++ stylesheets/utilities.scss
@@ -13,3 +13,7 @@
 .rotate-180 {
   transform: rotateX(180deg);
 }
+
+.min-h-screen {
+  min-height: 100vh;
+}
